remove UnclosedTsFileResource
