+++ server/src/test/java/org/apache/iotdb/db/engine/storagegroup/StorageGroupProcessorTest.java
@@ -138,7 +138,7 @@ public class StorageGroupProcessorTest {
 
     Assert.assertEquals(1, tsfileResourcesForQuery.size());
 
-    UnclosedTsFileResource tsfileResource = (UnclosedTsFileResource) tsfileResourcesForQuery.get(0);
+    TsFileResource tsfileResource = tsfileResourcesForQuery.get(0);
     Assert.assertEquals(
         0, tsfileResource.getChunkMetadataList(new PartialPath(deviceId, measurementId)).size());
     List<ReadOnlyMemChunk> memChunks =
